+++ ReferenceTests/Src/Catch2/Catch_Testset02/UT_Tests04.cpp
@@ -28,96 +28,42 @@ Notes: None
 
 namespace CatchTestset02
 {
-    TEST_CASE( "Testset02::Tests04. cout", "[cout]" )
+    TEST_CASE( "Testset02::Tests04. Testname", "[cout]" )
     {
-        std::cout << "cout 01\n";
-        INFO("Info 01");
-        std::cout << "cout 02\n";
-        CHECK( 1 == 2 );
-        std::cout << "cout 03\n";
+        int x = 42;
+        int y = 42;
 
-        SECTION("Root Section 1")
-        {
-            std::cout << "cout 04a\ncout 04b\n";
-            CHECK( 13 == 14 );
-        }
+        CHECK( x == y );
 
-        INFO("Info 04");
-        SECTION("Root Section 2")
-        {
-            INFO("Info 05");
-            std::cout << "cout 05a\ncout 05b\n";
-            CHECK( 23 == 24 );
-        }
+        WARN("Test case name only differes in case.");
     }
 
-    TEST_CASE( "Testset02::Tests04. cerr", "[cerr]" )
+    TEST_CASE( "Testset02::Tests04. TestName", "[cout]" )
     {
-        std::cerr << "cerr 01\n";
-        INFO("Info 01");
-        std::cerr << "cerr 02\n";
-        CHECK( 1 == 2 );
-        std::cerr << "cerr 03\n";
+        int x = 42;
+        int y = 47;
 
-        SECTION("Root Section 1")
-        {
-            std::cerr << "cerr 04a\ncerr 04b\n";
-            CHECK( 13 == 14 );
-        }
+        CHECK( x == y );
 
-        INFO("Info 04");
-        SECTION("Root Section 2")
-        {
-            INFO("Info 05");
-            std::cerr << "cerr 05a\ncerr 05b\n";
-            CHECK( 23 == 24 );
-        }
+        WARN("Test case name only differes in case.");
     }
 
-    TEST_CASE( "Testset02::Tests04. cout cerr", "[cerr][cout]" )
+    TEST_CASE( "Testset02::Tests04. TESTNAME", "[cout]" )
     {
-        std::cerr << "cerr 01\n";
-        INFO("Info 01");
-        std::cout << "cout 01\n";
-        CHECK( 1 == 2 );
-        std::cerr << "cerr 02\n";
+        int x = 42;
+        int y = 47;
 
-        SECTION("Root Section 1")
-        {
-            std::cout << "cout 02a\ncout 02b\n";
-            CHECK( 13 == 14 );
-        }
-
-        INFO("Info 04");
-        SECTION("Root Section 2")
-        {
-            INFO("Info 05");
-            std::cerr << "cerr 03ea\ncerr 03b\n";
-            CHECK( 23 == 24 );
-        }
+        CHECK( x == y );
+        CHECK( x >= y );
     }
 
-    TEST_CASE("Testset02::Tests04. cout cerr no failures", "[cerr][cout]")
+    TEST_CASE( "Testset02::Tests04. testname", "[cout]" )
     {
-        std::cerr << "cerr 01\n";
-        INFO( "Info 01" );
-        std::cout << "cout 01\n";
-        CHECK( 1 == 1 );
-        std::cerr << "cerr 02\n";
+        int x = 42;
+        int y = 42;
 
-        SECTION("Root Section 1")
-        {
-            std::cout << "cout 02a\ncout 02b\n";
-            CHECK( 14 == 14 );
-        }
-
-        INFO("Info 04");
-        SECTION("Root Section 2")
-        {
-            INFO("Info 05");
-            std::cerr << "cerr 03a\ncerr 03b\n";
-            CHECK( 24 == 24 );
-        }
+        CHECK( x == y );
+        CHECK( x >= y );
     }
 
 } // End namespace: CatchTestset02